requirements.txt
streamlit~=0.78.0
nibabel~=3.2.1
numpy~=1.19.5
matplotlib~=3.3.4
imageio~=2.9.0
scikit-image~=0.18.1
tensorflow-gpu~=2.4.1
scipy~=1.6.1
nilearn~=0.7.0